refactor(Assign.Vue): Change the layout to flex
I change the original three divs for class="left",class="center"and
class="right" to three el-col element, and delete corresponding of three
original divs which are not necessary. I add a new button to make left
el-col flexible and the center el-col can adapt to changes throughout
the window.